refactor: move block-arrival orchestration into node/chainman
Move ProcessBlock, AskForOutstandingBlocks and GridcoinServices from
main.cpp into node/chainman.{h,cpp}, next to ReorganizeChain and
SetBestChain -- their main collaborators (issue #3125, workstream C1).
Pure code move, no behavior change.

- chainman.cpp drops its extern forward declarations of
  GridcoinServices/AskForOutstandingBlocks (now declared in chainman.h)
  and gains the includes the moved code needs (chainparams, checkpoints,
  net, node/orphan_blocks, gridcoin/staking/spam) plus ad-hoc externs
  for g_seen_stakes and g_v11_timestamp, whose definitions stay in
  main.cpp to limit churn (same pattern as node/orphan_blocks.cpp and
  node/coherence.cpp).
- Callers include node/chainman.h directly (miner.cpp,
  net_processing.cpp, node/blockstorage.cpp; rpc/blockchain.cpp already
  did and loses its local AskForOutstandingBlocks declaration), matching
  how the A4 chainman extraction handled its callers -- a main.h
  re-include would introduce a main <-> node/chainman module cycle.

main.cpp -205 lines.
